Only use and update a single comment (#20)
Rather than creating a new comment for each commit, this makes it so only one comment will be created for each PR and it will be updated for each commit.

While there's a definite tradeoff here as it can be nice to see the history, the noise created by multiple comments just isn't worth it most of the time.